requirements.txt
spotipy>=2.22.1
python-dotenv>=1.0.0
RPLCD>=1.3.4
RPi.GPIO>=0.7.1